separation of tenant_repository and tenant_service decorators are moved to separate folder
