+++ backend/dataall/core/permissions/api/resolvers.py
@@ -4,13 +4,14 @@ import os
 from dataall.base.aws.sts import SessionHelper
 from dataall.base.aws.parameter_store import ParameterStoreManager
 from dataall.core.permissions.db.tenant.tenant_policy_repositories import TenantPolicy
+from dataall.core.permissions.services.tenant_policy_service import TenantPolicyService
 
 log = logging.getLogger(__name__)
 
 
 def update_group_permissions(context, source, input=None):
     with context.engine.scoped_session() as session:
-        return TenantPolicy.update_group_permissions(
+        return TenantPolicyService.update_group_permissions(
             session=session,
             username=context.username,
             groups=context.groups,
@@ -22,18 +23,15 @@ def update_group_permissions(context, source, input=None):
 
 def list_tenant_permissions(context, source):
     with context.engine.scoped_session() as session:
-        return TenantPolicy.list_tenant_permissions(session=session, username=context.username, groups=context.groups)
+        return TenantPolicyService.list_tenant_permissions(
+            session=session, username=context.username, groups=context.groups
+        )
 
 
 def list_tenant_groups(context, source, filter=None):
     with context.engine.scoped_session() as session:
-        return TenantPolicy.list_tenant_groups(
-            session=session,
-            username=context.username,
-            groups=context.groups,
-            uri=None,
-            data=filter if filter else {},
-            check_perm=True,
+        return TenantPolicyService.list_tenant_groups(
+            session=session, username=context.username, groups=context.groups, data=filter if filter else {}
         )
