+++ backend/dataall/core/permissions/db/tenant/tenant_models.py
@@ -5,7 +5,7 @@ from sqlalchemy.orm import relationship
 
 from dataall.base.db import Base, utils
 
-from  dataall.core.permissions.db.permission.permission_models import Permission
+from dataall.core.permissions.db.permission.permission_models import Permission
 
 
 class TenantPolicy(Base):
