+++ backend/local_graphql_server.py
@@ -10,7 +10,8 @@ from dataall.base.api import get_executable_schema
 from dataall.core.tasks.service_handlers import Worker
 from dataall.core.permissions.constants import permissions
 from dataall.core.permissions.db import save_permissions_with_tenant
-from dataall.core.permissions.db.tenant.tenant_policy_repositories import TenantPolicy
+from dataall.core.permissions.services.tenant_policy_service import TenantPolicyService
+
 from dataall.base.db import get_engine, Base
 from dataall.base.searchproxy import connect, run_query
 from dataall.base.loader import load_modules, ImportMode
@@ -72,7 +73,7 @@ def request_context(headers, mock=False):
 
     for group in groups:
         with engine.scoped_session() as session:
-            TenantPolicy.attach_group_tenant_policy(
+            TenantPolicyService.attach_group_tenant_policy(
                 session=session,
                 group=group,
                 permissions=permissions.TENANT_ALL,
